wifi-users/requirements.txt
flask~=1.1.2
python-dotenv~=0.15.0
flask_sqlalchemy
pandas~=1.1.5
flask-wtf
WTForms~=2.3.3
pip~=20.0.2
pytz~=2020.4
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
numpy~=1.19.4
setuptools~=44.0.0
MarkupSafe~=1.1.1
python-dateutil~=2.8.1
SQLAlchemy~=1.3.20
six~=1.15.0